fix: Tests, most raicing conditions fixed
