+++ deployments/deployed-contracts-hardhat.json
@@ -45,37 +45,37 @@
     "address": "0xc831fAe49f2c324704F7133CB8bE08BC768904cf"
   },
   "DebtMarketImpl": {
-    "address": "0x53933349dA8E97b77c1f43Ba01192adb8C510fA7"
+    "address": "0x51998F16F707a0cdd5ECE2a56c034552dF3fb855"
   },
   "DebtMarket": {
     "address": "0xAAd52A86F916f46c2Dc87f2BfCEA0B03aEAa8AE1"
   },
   "GenericLogic": {
-    "address": "0xF340913980c21cb9D96bE990451c5471FABBD2B9",
-    "deployer": "0x5b69E6884C70f42819Fb35Bf3C25578ee11AAA15"
-  },
-  "ReserveLogic": {
-    "address": "0x0A66e11653e32D082F3330Dc99687ce42Eb53121",
-    "deployer": "0x5b69E6884C70f42819Fb35Bf3C25578ee11AAA15"
-  },
-  "NftLogic": {
     "address": "0xD5A42dEB39C8EfbCE0f81cEb4c51500501809e89",
     "deployer": "0x5b69E6884C70f42819Fb35Bf3C25578ee11AAA15"
   },
-  "ValidationLogic": {
+  "ReserveLogic": {
     "address": "0x4E76d6429C6f9dD22F87f10BcA800664B8Cc118b",
     "deployer": "0x5b69E6884C70f42819Fb35Bf3C25578ee11AAA15"
   },
-  "SupplyLogic": {
+  "NftLogic": {
     "address": "0x153a75c692b6f047974bc96e30d20546B0Bc2992",
     "deployer": "0x5b69E6884C70f42819Fb35Bf3C25578ee11AAA15"
   },
-  "BorrowLogic": {
+  "ValidationLogic": {
     "address": "0x17624D935b72A4d765d7A69a68E7C0725A38d5bc",
     "deployer": "0x5b69E6884C70f42819Fb35Bf3C25578ee11AAA15"
   },
-  "LiquidateLogic": {
+  "SupplyLogic": {
     "address": "0x7ECdCa35F3B55acE694A7DA48a759444Fe81b943",
+    "deployer": "0x5b69E6884C70f42819Fb35Bf3C25578ee11AAA15"
+  },
+  "BorrowLogic": {
+    "address": "0xa25a322c846E3be3207365390A02864fdEDfA0CD",
+    "deployer": "0x5b69E6884C70f42819Fb35Bf3C25578ee11AAA15"
+  },
+  "LiquidateLogic": {
+    "address": "0x7dDc7b2963AC96336Ec238e816FCbc30aDB113AD",
     "deployer": "0x5b69E6884C70f42819Fb35Bf3C25578ee11AAA15"
   },
   "LiquidateMarketsLogic": {
@@ -83,114 +83,115 @@
     "deployer": "0x5b69E6884C70f42819Fb35Bf3C25578ee11AAA15"
   },
   "LendingLogic": {
-    "address": "0xa25a322c846E3be3207365390A02864fdEDfA0CD",
+    "address": "0x58B6e24b2a462Ef81e154515Ab7adc804512325c",
     "deployer": "0x5b69E6884C70f42819Fb35Bf3C25578ee11AAA15"
   },
   "ConfiguratorLogic": {
-    "address": "0x7dDc7b2963AC96336Ec238e816FCbc30aDB113AD",
+    "address": "0xAAF09660FafD2138fC9e5dA036D2a101A623b52F",
     "deployer": "0x5b69E6884C70f42819Fb35Bf3C25578ee11AAA15"
   },
   "LendPoolImpl": {
-    "address": "0x58B6e24b2a462Ef81e154515Ab7adc804512325c"
+    "address": "0x222193c6A20b1596814ffEea6B320cbF3cFe2c6b"
   },
   "LendPool": {
-    "address": "0x80Fa1fA8B0CaF15CFC3961C934cFA860945d9139"
+    "address": "0x222193c6A20b1596814ffEea6B320cbF3cFe2c6b",
+    "deployer": "0x5b69E6884C70f42819Fb35Bf3C25578ee11AAA15"
   },
   "LendPoolLoanImpl": {
-    "address": "0x5c769753F192449c3c6AeC7BF2145928eb261699"
+    "address": "0xDB8c2AE619685b26C26a9ff48698049C363784f3"
   },
   "LendPoolLoan": {
     "address": "0x7D2386aA0b699eD5C9a8D9caa2efDA648C2418e7"
   },
   "LendPoolConfiguratorImpl": {
-    "address": "0xDB8c2AE619685b26C26a9ff48698049C363784f3"
+    "address": "0x32727fA2ad74483C8DC1F899C5c06AcCd5eDAED8"
   },
   "LendPoolConfigurator": {
     "address": "0x8E170f8919F9f6e58EEbd6E7bF0f1E2CFfaBe5E4"
   },
   "UToken": {
-    "address": "0xBb94c69258B09B351E31972aDd0951f30E7f5ed6",
+    "address": "0x8286b7d42cB141E1871a3475bB8E0C3F84db0352",
     "deployer": "0x5b69E6884C70f42819Fb35Bf3C25578ee11AAA15"
   },
   "DebtToken": {
-    "address": "0x76F965B99e9E216a648A8dB9F8aB459e880181Ff",
+    "address": "0x53933349dA8E97b77c1f43Ba01192adb8C510fA7",
     "deployer": "0x5b69E6884C70f42819Fb35Bf3C25578ee11AAA15"
   },
   "ReserveOracleImpl": {
-    "address": "0x8e090D5B023252bE8d05d4c33b959A6F4A8BdD9e"
+    "address": "0xf3D4979e1d7b8a8b9750e79ceF466Efc46298403"
   },
   "ReserveOracle": {
-    "address": "0x3Bd92b2536c4F1537c7c95eA4a203c182AC57973",
+    "address": "0x941C6c63Def537699014d15605828339475ed33c",
     "deployer": "0x5b69E6884C70f42819Fb35Bf3C25578ee11AAA15"
   },
   "MockReserveOracle": {
-    "address": "0xd81cfDb25C4a27795c17A893F7813590614f2A90",
-    "deployer": "0x5b69E6884C70f42819Fb35Bf3C25578ee11AAA15"
-  },
-  "MockChainlinkOracle": {
     "address": "0x06A6737A0713e3DAe8DEfA7565f89D835C1594B7",
     "deployer": "0x5b69E6884C70f42819Fb35Bf3C25578ee11AAA15"
   },
-  "NFTOracleImpl": {
-    "address": "0xc27F09fe69Ddd42b2d84e15eC384B5462339727A"
-  },
-  "NFTOracle": {
+  "MockChainlinkOracle": {
     "address": "0x9718F7A661C0137d13EB4f0659dA5E8697075492",
     "deployer": "0x5b69E6884C70f42819Fb35Bf3C25578ee11AAA15"
   },
-  "MockNFTOracle": {
-    "address": "0x5cb48eC1643Fd278332bBB46090126bFCBC5b616",
+  "NFTOracleImpl": {
+    "address": "0x0c6ee9D95dD6e39C1808d3fF219a6cE9b70E8dA7"
+  },
+  "NFTOracle": {
+    "address": "0xc5b75fA87Bc8F55bD989642DFE0Bbdd44748cD3b",
     "deployer": "0x5b69E6884C70f42819Fb35Bf3C25578ee11AAA15"
   },
-  "InterestRate": {
+  "MockNFTOracle": {
     "address": "0x2bDc03E97c90C269099931986078fb9e353a65C6",
     "deployer": "0x5b69E6884C70f42819Fb35Bf3C25578ee11AAA15"
   },
+  "InterestRate": {
+    "address": "0x24A7d384dea2e1Dea6d2fE4f04A3197b2d7a12de",
+    "deployer": "0x5b69E6884C70f42819Fb35Bf3C25578ee11AAA15"
+  },
   "rateStrategyWETH": {
-    "address": "0x2bDc03E97c90C269099931986078fb9e353a65C6"
+    "address": "0x495DeE8247914e3176a4F9767596d15a09230aC7"
   },
   "WETHGatewayImpl": {
-    "address": "0x58954c33A07348dB1ad3d9758ddE1a78223a2220"
+    "address": "0x6e41fB61F92789d78ed6b7B70B5452c13706c2fE"
   },
   "WETHGateway": {
-    "address": "0x9a195074E9932098bb90f5a80E3Bdf87Ed9244Ac",
+    "address": "0xD0330Ac56A182aC22864D67D6f6E76aD628c9b2e",
     "deployer": "0x5b69E6884C70f42819Fb35Bf3C25578ee11AAA15"
   },
   "PunkGatewayImpl": {
-    "address": "0x81B8C7e1cC7037940893dD3320562E950E98848c"
+    "address": "0x51bda59320165617F7209a38D08b11ccefF58F1a"
   },
   "PunkGateway": {
-    "address": "0x7326573C3689831ADBf8050F1e5a2fB31C8441d8",
+    "address": "0x816b2d94590fbcE4C06a70FBb770323166867549",
     "deployer": "0x5b69E6884C70f42819Fb35Bf3C25578ee11AAA15"
   },
   "ReservoirAdapterImpl": {
-    "address": "0xD024caf0B8667D0D50e87C35Da1AF295ccd96CeC"
+    "address": "0x35793856D2b49B365748b532149E1C52Cfa1744F"
   },
   "ReservoirAdapter": {
     "address": "0xa62929737dd645647f2030b1D0074F9FB687051F"
   },
   "WalletBalanceProvider": {
-    "address": "0x35793856D2b49B365748b532149E1C52Cfa1744F",
-    "deployer": "0x5b69E6884C70f42819Fb35Bf3C25578ee11AAA15"
-  },
-  "UnlockdProtocolDataProvider": {
     "address": "0x8Fe994bA879dF436bDA80d1fc559d192608939f6",
     "deployer": "0x5b69E6884C70f42819Fb35Bf3C25578ee11AAA15"
   },
-  "UIPoolDataProvider": {
+  "UnlockdProtocolDataProvider": {
     "address": "0xF79a1ab14b401d7032bdDD20D8Eb061557fa8D87",
     "deployer": "0x5b69E6884C70f42819Fb35Bf3C25578ee11AAA15"
   },
+  "UIPoolDataProvider": {
+    "address": "0xF07191639A5ca3d5D28206429Df3E3adCc020b3f",
+    "deployer": "0x5b69E6884C70f42819Fb35Bf3C25578ee11AAA15"
+  },
   "SelfdestructTransferMock": {
-    "address": "0x034226F4d7C94427dbDa9e09ad87A8597ECb7325",
+    "address": "0x5aC94dED0Fe79C890AF99f92164D93941a3F6Ac7",
     "deployer": "0x5b69E6884C70f42819Fb35Bf3C25578ee11AAA15"
   },
   "MockYVaultImpl": {
-    "address": "0x7D509b64f31Dcc7bc7848F9789F1e99c5e0f5C22",
+    "address": "0x24A7d384dea2e1Dea6d2fE4f04A3197b2d7a12de",
     "deployer": "0x5b69E6884C70f42819Fb35Bf3C25578ee11AAA15"
   },
   "MockYVault": {
-    "address": "0xC8bcB2Fd3EA4D38BC62663Cf782805724eFa6Bb0",
+    "address": "0x222193c6A20b1596814ffEea6B320cbF3cFe2c6b",
     "deployer": "0x5b69E6884C70f42819Fb35Bf3C25578ee11AAA15"
   }
 }
